+++ package.json
@@ -10,6 +10,7 @@
     "luxon": "^3.5.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^7.3.0",
     "tailwindcss": "^4.0.11"
   },
   "scripts": {
